mise à jour

+++ src/main/java/com/springjava/javaspring/controller/CompetenceController.java
@@ -1,13 +1,14 @@
 package com.springjava.javaspring.controller;
-
-import com.fasterxml.jackson.annotation.JsonView;
 import com.springjava.javaspring.dao.CompetenceDao;
 import com.springjava.javaspring.model.Competence;
-import com.springjava.javaspring.view.JsonViewCustom;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @CrossOrigin
@@ -16,39 +17,42 @@ public class CompetenceController {
 
     CompetenceDao competenceDao;
 
-     @Autowired
-     CompetenceController(CompetenceDao competenceDao){
-         this.competenceDao = competenceDao;
+    @Autowired
+    CompetenceController(CompetenceDao competenceDao){
+        this.competenceDao = competenceDao;
     }
 
-    @GetMapping("/competence/{id}")
-    @JsonView({JsonViewCustom.VueCompetence.class})
-    public Competence getCompetences(@PathVariable int id){
+    @GetMapping("/user/competence/{id}")
+    public ResponseEntity<Competence> getCompetences(@PathVariable int id){
 
-        return competenceDao.findById(id).orElse(null);
+        Optional<Competence> competence = competenceDao.findById(id);
+
+        return competence.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.noContent().build());
     }
 
-    @GetMapping("/competences")
-    @JsonView({JsonViewCustom.VueCompetence.class})
-    public List<Competence> getCompetences(){
+    @GetMapping("/user/competences")
+    public ResponseEntity<List<Competence>> getCompetences(){
 
-        return competenceDao.findAll();
+        return ResponseEntity.ok(competenceDao.findAll());
     }
 
+    @PostMapping("/admin/competence")
+    public ResponseEntity<String> addCompetence (@RequestBody Competence competence) {
 
-    @PostMapping("/competence")
-    public boolean addCompetence (@RequestBody Competence competence) {
+        competence =competenceDao.saveAndFlush(competence);
 
-         competenceDao.save(competence);
-
-        return true;
+        return ResponseEntity.created(URI.create("/user/competence/" + competence.getId())).build();
     }
 
     @DeleteMapping("/competence/{id}")
-    public boolean deleteCompetence (@PathVariable int id) {
+    public ResponseEntity<Integer> deleteCompetence (@PathVariable int id) {
 
-         competenceDao.deleteById(id);
-        return true;
+        if(competenceDao.existsById(id)){
+            competenceDao.deleteById(id);
+            return ResponseEntity.ok(id);
+        }else{
+            return ResponseEntity.noContent().build() ;
+        }
     }
